detangle

+++ utils.pde
@@ -1,4 +1,4 @@
-float random_grid_x(){
+int random_grid_x(){
   return OFFSET +  ( int(random(0,width))/GRID_X )*GRID_X ;
 }
